fixed remaining issue with corruption on startup
The corruption on startup was due to the serial write happening too fast. the baud rate is only 1200, which is ok for C64, but not for the basic program, or the wiring I had setup. adding a small delay between sending chanracters, similar to when you type normally  fixed the fault